+++ packages/api/resolvers/mutations/logout.js
@@ -6,7 +6,10 @@ const logger = console;
 export default async function logout(root, { token }, context) {
   log('mutation logout');
   try {
-    await accountsServer.logout({ token, userId: context.userId });
+    await accountsServer.logout({
+      token: token || context.loginToken,
+      userId: context.userId,
+    });
     return { success: true };
   } catch (e) {
     logger.error(e);