+++ changelog.md
@@ -3,6 +3,7 @@
 This is a bugfix release based on learnings from upgrading client projects
 
 ## Breaking changes
+- [api] Fix `logout` regression with not falling back to the current token when used without explicit token
 - [core] `Users.createUser` now sends messages by default, you have to explicitly bail out by either providing an option `skipMessaging` to true or set the new `autoMessagingAfterUserCreation` module param of users to `false`. Take a look at your seed scripts.
 - [platform] Certain fields like `emails` and `services` are now blocked from passing to `Users.createUser`. If you have used `emails` in `createUser` seeding before, use email. If you want to skip e-mail verification forcefully, use `initialPassword: true`. See the changes in the minimal example seed file to get a glimpse.